rename(design-system): move Box component to primitives
