+++ apps/web/components/ControlPanel/types.ts
@@ -1,4 +1,4 @@
-import { ComponentPropsByBox } from '@pigyuma/design-system/patterns';
+import { ComponentPropsByBox } from '@pigyuma/design-system/primitives';
 import Panel from '../Panel';
 
 export type ControlPanelProps = Omit<ComponentPropsByBox<typeof Panel>, 'role' | 'placement'>;
