+++ apps/web/components/LayerTreeSection/LayerTreeSection.tsx
@@ -1,5 +1,5 @@
 import mixins from '@pigyuma/design-system/mixins';
-import { Box } from '@pigyuma/design-system/patterns';
+import { Box } from '@pigyuma/design-system/primitives';
 import { useUIData } from '@pigyuma/ui-design-tool';
 import clsx from 'clsx';
 import React from 'react';
