+++ apps/web/components/Workspace/types.ts
@@ -1,4 +1,4 @@
-import { ComponentPropsByBox, ComponentRefByBox } from '@pigyuma/design-system/patterns';
+import { ComponentPropsByBox, ComponentRefByBox } from '@pigyuma/design-system/primitives';
 import { UIDesignCanvasProps } from '@pigyuma/ui-design-tool';
 
 export type WorkspaceProps = ComponentPropsByBox<'div'> & UIDesignCanvasProps;
