+++ packages/design-system/components/Button/types.ts
@@ -1,4 +1,4 @@
-import { ComponentPropsByBox, ComponentRefByBox } from '@/patterns/Box';
+import { ComponentPropsByBox, ComponentRefByBox } from '@/primitives/Box';
 
 export type ButtonProps = ComponentPropsByBox<'button'>;
 export type ButtonRef = ComponentRefByBox<'button'>;
