+++ packages/design-system/patterns/index.ts
@@ -1,2 +1,1 @@
-export * from './Box';
-export { default as Box } from './Box';
+export const _ = undefined;
