+++ packages/design-system/primitives/index.ts
@@ -1,1 +1,2 @@
-export const _ = undefined;
+export * from './Box';
+export { default as Box } from './Box';
